+++ src/main/java/kr/jay/okrver3/interfaces/feedback/FeedbackDtoMapper.java
@@ -6,12 +6,5 @@ import kr.jay.okrver3.interfaces.feedback.request.FeedbackSaveRequest;
 
 @Component
 public class FeedbackDtoMapper {
-	public FeedbackSaveCommand of(FeedbackSaveRequest dto) {
-		return new FeedbackSaveCommand(
-			dto.opinion(),
-			dto.grade(),
-			dto.projectToken(),
-			dto.initiativeToken()
-		);
-	}
+
 }
