+++ src/main/java/kr/jay/okrver3/interfaces/project/ProjectApiController.java
@@ -27,6 +27,7 @@ import kr.jay.okrver3.domain.project.SortType;
 import kr.jay.okrver3.domain.project.service.command.ProjectDetailRetrieveCommand;
 import kr.jay.okrver3.domain.project.service.info.ProjectSideMenuInfo;
 import kr.jay.okrver3.domain.user.User;
+import kr.jay.okrver3.interfaces.feedback.request.FeedbackSaveRequest;
 import kr.jay.okrver3.interfaces.project.request.ProjectInitiativeSaveRequest;
 import kr.jay.okrver3.interfaces.project.request.ProjectKeyResultSaveRequest;
 import kr.jay.okrver3.interfaces.project.request.ProjectSaveRequest;
@@ -200,6 +201,20 @@ public class ProjectApiController {
 		);
 
 	}
+
+	@PostMapping("/feedback")
+	public ResponseEntity<String> registerFeedback(
+		@RequestBody @Valid FeedbackSaveRequest requestDto,
+		Authentication authentication) {
+
+		return Response.successOk(
+			projectFacade.registerFeedback(
+				mapper.of(requestDto),
+				getUserFromAuthentication(authentication)
+			)
+		);
+	}
+
 
 	private User getUserFromAuthentication(Authentication authentication) {
 		return ClassUtils.getSafeCastInstance(authentication.getPrincipal(), User.class)
