+++ src/test/java/kr/jay/okrver3/application/project/ProjectFacadeTest.java
@@ -51,6 +51,7 @@ import kr.jay.okrver3.infrastructure.initiative.InitiativeQueryDslRepository;
 import kr.jay.okrver3.infrastructure.notification.NotificationJDBCRepository;
 import kr.jay.okrver3.infrastructure.project.ProjectQueryDslRepository;
 import kr.jay.okrver3.infrastructure.project.ProjectRepositoryImpl;
+import kr.jay.okrver3.interfaces.feedback.FeedbackSaveCommand;
 
 @DataJpaTest
 @Import({ProjectFacade.class, ProjectServiceImpl.class, UserServiceImpl.class,
@@ -363,6 +364,24 @@ class ProjectFacadeTest {
 			assertThat(content.get(i).initiativeToken()).isEqualTo(savedInitiativeTokenRecentlyCreatedOrder.get(i));
 		}
 
+	}
+
+	@Test
+	@Sql("classpath:insert-project-date.sql")
+	void 팀원의_행동전략에_피드백을_추가하면_기대하는_응답을_리턴한다() throws Exception {
+
+		FeedbackSaveCommand command =
+			new FeedbackSaveCommand("피드백 작성", "GOOD_IDEA", "mst_Kiwqnp1Nq6lb6421",
+				"ini_ixYjj5aaafeab3AH8");
+
+		String response =
+			sut.registerFeedback(
+				command,
+				getUser(3L)
+			);
+
+		assertThat(response).containsPattern(
+			Pattern.compile("initiative-[a-zA-Z0-9]{9}"));
 	}
 
 }
