+++ src/test/java/kr/jay/okrver3/interfaces/project/ProjectApiControllerAcceptanceTest.java
@@ -39,6 +39,7 @@ import kr.jay.okrver3.common.exception.ErrorCode;
 import kr.jay.okrver3.common.utils.JwtTokenUtils;
 import kr.jay.okrver3.domain.keyresult.KeyResult;
 import kr.jay.okrver3.domain.project.Project;
+import kr.jay.okrver3.interfaces.feedback.request.FeedbackSaveRequest;
 import kr.jay.okrver3.interfaces.project.request.ProjectInitiativeSaveRequest;
 import kr.jay.okrver3.interfaces.project.request.ProjectKeyResultSaveRequest;
 import kr.jay.okrver3.interfaces.project.request.ProjectSaveRequest;
@@ -509,6 +510,28 @@ public class ProjectApiControllerAcceptanceTest {
 			then()
 			.statusCode(HttpStatus.OK.value())
 			.extract().body().asString();
+
+	}
+
+	@Test
+	void 팀원의_행동전략에_피드백을_추가하면_기대하는_응답을_리턴한다() throws Exception {
+
+		final String response = RestAssured.
+
+			given()
+			.header("Authorization", "Bearer " + authToken)
+			.contentType(ContentType.JSON)
+			.body(new FeedbackSaveRequest("피드백 작성", "GOOD_IDEA", "project-fgFHxGWeIUQt", "ini_ixYjj5nODqtb3AH8")).
+
+			when()
+			.post(baseUrl).
+
+			then()
+			.statusCode(HttpStatus.CREATED.value())
+			.extract().body().asString();
+
+		assertThat(response).containsPattern(
+			Pattern.compile("feedback-[a-zA-Z0-9]{11}"));
 
 	}
 }
