+++ src/test/java/kr/jay/okrver3/interfaces/project/ProjectApiControllerTest.java
@@ -26,6 +26,7 @@ import kr.jay.okrver3.common.exception.ErrorCode;
 import kr.jay.okrver3.common.exception.OkrApplicationException;
 import kr.jay.okrver3.domain.project.ProjectType;
 import kr.jay.okrver3.domain.user.User;
+import kr.jay.okrver3.interfaces.feedback.request.FeedbackSaveRequest;
 import kr.jay.okrver3.interfaces.project.request.ProjectInitiativeSaveRequest;
 import kr.jay.okrver3.interfaces.project.request.ProjectKeyResultSaveRequest;
 import kr.jay.okrver3.interfaces.project.request.ProjectSaveRequest;
@@ -298,6 +299,26 @@ class ProjectApiControllerTest {
 		}
 
 	}
+
+	@Test
+	@Sql("classpath:insert-project-date.sql")
+	void 팀원의_행동전략에_피드백을_추가하면_기대하는_응답을_리턴한다() throws Exception {
+
+		FeedbackSaveRequest requestDto =
+			new FeedbackSaveRequest("피드백 작성", "GOOD_IDEA", "mst_Kiwqnp1Nq6lb6421",
+				"ini_ixYjj5aaafeab3AH8");
+
+		ResponseEntity<String> response =
+			sut.registerFeedback(
+				requestDto,
+				getAuthenticationToken(3L)
+			);
+
+		assertThat(response.getBody()).containsPattern(
+			Pattern.compile("initiative-[a-zA-Z0-9]{9}"));
+	}
+
+
 
 	private UsernamePasswordAuthenticationToken getAuthenticationToken(long value) {
 		User user = em.createQuery("select u from User u where u.id = :userSeq", User.class)
